Added Openings and BoxTrigger to room and worked on the room spawning
The adjacent room spawning is not so easy, because I need to find a fitting room by choosing one by random, get a random entrance from that room, rotate the room and check if it fits the position, if so I need to spawn it and check, which openings are connected now or not and mark the right ones as connected. Then I add the new room to the room list of the dungeon generator.