Got generation working, but did discover bugs.
First bug, sometimes the generator gets stuck because it spawned a room where it can't place another one on one of the entrances. So I build a fitting peace that would solve the problem, without looking bad.
But with that I discovered another bug, that the rotation of the collider for checking or/and the rotation for the check is not correct. So I need to fix that.